requirements.txt
mysql-connector-python==8.0.16
Flask==1.0.2
boto==2.49.0
boto3==1.9.137
botocore==1.12.137